refactor: refactor services and dao files
refactor services and dao files
