+++ __test__/routes/user_show_me_route.test.ts
@@ -4,9 +4,9 @@ import bcrypt from 'bcryptjs';
 
 import config from '../../src/config/app';
 import server from '../../src/server/http_server';
-import getUser from '../../src/dao/users/user_get_one_service';
-import createUser from '../../src/dao/users/user_create_service';
-import deleteUser from '../../src/dao/users/user_physical_delete_service';
+import getUser from '../../src/dao/users/user_get_one_dao';
+import createUser from '../../src/dao/users/user_create_dao';
+import deleteUser from '../../src/dao/users/user_delete_dao';
 import globalApiPath from '../../src/utils/global_api_path/global_api_path';
 
 const apiPath = globalApiPath();
