--- src/dao/commons/api_root_service.ts
@@ -1,8 +1,0 @@
-import config from '@config/app';
-
-export default () => {
-    const baseApiUrl = config.api.prefix;
-    const route = '/' + baseApiUrl + '/info';
-
-    return { success: true, data: route, error: null };
-};
