+++ src/services/cliente/user_auth/login_service.ts
@@ -1,5 +1,5 @@
 import httpMsg from '@utils/http_messages/http_msg';
-import servFindOneUser from '@dao/users/user_get_one_service';
+import servFindOneUser from '@dao/users/user_get_one_dao';
 import servCheckPassword from '@functions/check_password';
 import servGenerateToken from '@functions/generate_token_access';
