+++ src/services/cliente/user_auth/register_confirm_service.ts
@@ -1,7 +1,7 @@
 import config from '@config/app/index';
 import httpMsg from '@utils/http_messages/http_msg';
-import servUpdateUser from '@dao/users/user_update_service';
-import servFindOneUser from '@dao/users/user_get_one_service';
+import servUpdateUser from '@dao/users/user_update_dao';
+import servFindOneUser from '@dao/users/user_get_one_dao';
 // import sendEmail from '@utils/nodemailer/nodemailer/email_welcome';
 
 const errorCod = 'ERROR_USER_SIGNUP_CONFIRMATION';
