+++ src/services/cliente/user_auth/register_service.ts
@@ -2,9 +2,9 @@ import randtoken from 'rand-token';
 
 import config from '@config/app/index';
 import httpMsg from '@utils/http_messages/http_msg';
-import servCreateUser from '@dao/users/user_create_service';
-import servUpdateUser from '@dao/users/user_update_service';
-import servFindOneUser from '@dao/users/user_get_one_service';
+import servCreateUser from '@dao/users/user_create_dao';
+import servUpdateUser from '@dao/users/user_update_dao';
+import servFindOneUser from '@dao/users/user_get_one_dao';
 import servHashPassword from '@functions/generate_hash_password';
 import sendEmail from '@utils/nodemailer/nodemailer/email_verification';
