+++ src/services/cliente/user_me/index.ts
@@ -1,7 +1,7 @@
-import showAll from '../../admin/users/show_all_presenter';
-import showMe from './user_show_me_presenter';
-import updateMe from './user_update_me_presenter';
-import deleteMe from './user_delete_me_presenter';
+import showAll from '../../admin/users/show_all_service';
+import showMe from './user_show_me_service';
+import updateMe from './user_update_me_service';
+import deleteMe from './user_delete_me_service';
 
 export default {
     showAll,
