+++ src/services/cliente/user_me/user_delete_me_service.ts
@@ -1,8 +1,9 @@
 import httpMsg from '@utils/http_messages/http_msg';
-import servFindOneUser from '@dao/users/user_get_one_service';
+import servUpdateUser from '@dao/users/user_update_dao';
+import servFindOneUser from '@dao/users/user_get_one_dao';
 
-const errorCod = 'ERROR_USER_FIND_ME';
-const errorMsg = 'Failed to show user';
+const errorCod = 'ERROR_USER_DELETE_ME';
+const errorMsg = 'Failed to delete user';
 
 export default async (id: string) => {
     // Check required user data
@@ -12,7 +13,11 @@ export default async (id: string) => {
     const user = await getUser({ id, isDeleted: false, isRegistered: true });
     if (!user.success) return httpMsg.http422(user.error || '', errorCod);
 
-    return httpMsg.http200(user.data);
+    // Update the user
+    const updated = await updateUser(user.data.id, { isDeleted: true });
+    if (!updated.success) return httpMsg.http422(errorMsg, errorCod);
+
+    return httpMsg.http204(updated.data);
 };
 
 const checkRequiredDatas = (id: string) => /* istanbul ignore next */ {
@@ -36,4 +41,18 @@ const getUser = async (where: object) => {
     if (!result.success || !result.data) return { success: false, data: null, error: errorMsg };
 
     return { success: true, data: result.data, error: null };
+};
+
+const updateUser = async (id: string, datas: any) => {
+    const select = {
+        id: true,
+        isDeleted: true,
+    };
+
+    const result = await servUpdateUser(id, datas, select);
+
+    /* istanbul ignore if */
+    if (!result.success || !result.data) return { success: false, data: null };
+
+    return { success: true, data: result.data };
 };
