+++ src/services/cliente/user_me/user_update_me_service.ts
@@ -1,7 +1,7 @@
 import httpMsg from '@utils/http_messages/http_msg';
-import servUpdateUser from '@dao/users/user_update_service';
+import servUpdateUser from '@dao/users/user_update_dao';
 import servHashPassword from '@functions/generate_hash_password';
-import servFindOneUser from '@dao/users/user_get_one_service';
+import servFindOneUser from '@dao/users/user_get_one_dao';
 
 const errorCod = 'ERROR_USER_UPDATE_ME';
 const errorMsg = 'Failed to update user';
