+++ src/services/commons/api_info_service.ts
@@ -1,18 +1,18 @@
 import logger from '@utils/winston_file_logger/winston/logger';
 import httpMsg from '@utils/http_messages/http_msg';
-import apiInfo from '@dao/commons/api_info_service';
+import pkg from '@packagejson';
 
 const lblErr = 'API Info error';
 
 export default async () => {
     try {
-        const result = await apiInfo();
+        const data = {
+            name: pkg.name,
+            description: pkg.description,
+            version: pkg.version,
+        };
 
-        /* istanbul ignore if */
-        if (!result.success || !result.data)
-            return httpMsg.http422('Erro ao carregar informações', lblErr);
-
-        return httpMsg.http200(result.data);
+        return httpMsg.http200(data);
     } catch (err: any) /* istanbul ignore next */ {
         logger.error(`Erro ao carregar informações de API. ${err.message}`);
         return httpMsg.http422('Erro ao carregar informações', lblErr);
