+++ src/services/logs_infos/index.ts
@@ -1,4 +1,4 @@
-import listar from './listar_presenter';
+import listar from './listar_service';
 
 export default {
     listar,
